+++ src/main.js
@@ -4,7 +4,7 @@ import piniaPluginPersistedstate from 'pinia-plugin-persistedstate';
 import router from './router';
 import './style.css';
 import App from './App.vue';
-import '@/lib/api/attachAuthInterceptors';
+import '@/lib/api/http';
 import { useAuthStore } from '@/stores/auth';
 import '@fortawesome/fontawesome-free/css/all.min.css';
